+++ src/FSharpVSPowerTools/StandardCommandDispatcher.cs
@@ -24,7 +24,7 @@ namespace FSharpVSPowerTools
             GeneralOptionsPage generalOptions = (GeneralOptionsPage)(Package.GetGlobalService(typeof(GeneralOptionsPage)));
             if (!generalOptions.FormattingEnabled)
             {
-                Debug.WriteLine("Formatting feature is disabled in General option page.");
+                Debug.WriteLine("[Formatting] The feature is disabled in General option page.");
                 return;
             }