+++ RELEASE_NOTES.md
@@ -5,3 +5,7 @@
 
 #### 0.1.1 - February 5 2013
 * Add options to disable constituent parts
+
+#### 0.2.0 - Unreleased
+* Add navigation bar setting
+* Implement highlighting usage